+++ packages/esify/index.js
@@ -55,6 +55,9 @@ var TRANSFORMS = [
   {path: 'shopify-codemod/transforms/constant-function-expression-to-statement'},
   {path: 'shopify-codemod/transforms/global-reference-to-import'},
   {path: 'shopify-codemod/transforms/global-identifier-to-import'},
+  // Must appear after constant-function-expression-to-statement in order to remove
+  // unneeded semicolons from exported function declarations
+  {path: 'shopify-codemod/transforms/remove-empty-statements'},
 ];
 
 var OPTIONS = loadOptions();