+++ packages/shopify-codemod/transforms/remove-empty-statements.js
@@ -1,0 +1,6 @@
+export default function removeEmptyStatements({source}, {jscodeshift: j}, {printOptions = {quote: 'single'}}) {
+  return j(source)
+    .find(j.EmptyStatement)
+    .replaceWith()
+    .toSource(printOptions);
+}